Remove code to support EL6 (#202)
According to metadata.json EL6 is no longer supported so we can safely
remove the code that supported it.

This also makes the service idempotent on EL7 because the status command
was broken on systemd. Since we have a native unit file, we can rely on
the built in features.